fix(ops): address CI + CodeQL feedback (CHAOS-1630)
